pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-north-voila"
description = "A voila web application from Jupyter Notebook."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "NOMAD Authors", email = "fairmat@physik.hu-berlin.de" },
]
maintainers = [
    { name = "NOMAD Authors", email = "fairmat@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.4.1"
]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-north-voila"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "structlog",
    "mkdocs",
    "mkdocs-material==8.1.1",
    "pymdown-extensions",
    "mkdocs-click",
    "pytest-asyncio",
]
[dependency-groups]
north = [
    # According to the base image ensure compatibility of other dependencies with this version
    "jupyterlab==3.6.2",
    "ipywidgets",
    "voila",
    "ipyaggrid",
    "ipysheet",
    "ipydatagrid",
    "jupyter-flex",
]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415", # `import` should be at the top-level of a file
    "E501", # line too long
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["PLC0415"] # import-outside-top-level

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"

[project.entry-points.'nomad.plugin']
north_tool_entry_point = "nomad_north_voila.north_tools.voila_north_tool:north_tool_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
